move attactors to folder
