+++ calcit.cirru
@@ -1293,42 +1293,42 @@
                     |h $ %{} :Leaf (:at 1709570967583) (:by |rJG4IHzWf) (:text |fireworks)
                 |z5 $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1699553220765) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-aizawa")
+                    |T $ %{} :Leaf (:at 1715484559506) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/aizawa")
                     |b $ %{} :Leaf (:at 1709570977430) (:by |rJG4IHzWf) (:text |:as)
                     |h $ %{} :Leaf (:at 1709570974264) (:by |rJG4IHzWf) (:text |aizawa)
                 |z7 $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1699553232943) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-lorenz")
+                    |T $ %{} :Leaf (:at 1715484560986) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/lorenz")
                     |b $ %{} :Leaf (:at 1709570979733) (:by |rJG4IHzWf) (:text |:as)
                     |e $ %{} :Leaf (:at 1709570985482) (:by |rJG4IHzWf) (:text |lorenz)
                 |z8 $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1699553247355) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-fourwing")
+                    |T $ %{} :Leaf (:at 1715484562708) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/fourwing")
                     |b $ %{} :Leaf (:at 1709570987749) (:by |rJG4IHzWf) (:text |:as)
                     |e $ %{} :Leaf (:at 1709570991427) (:by |rJG4IHzWf) (:text |fourwing)
                 |z8T $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1706427309218) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-chen")
+                    |T $ %{} :Leaf (:at 1715484564109) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/chen")
                     |b $ %{} :Leaf (:at 1709570993351) (:by |rJG4IHzWf) (:text |:as)
                     |e $ %{} :Leaf (:at 1709570995295) (:by |rJG4IHzWf) (:text |chen)
                 |z8j $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1706459656210) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-sprott")
+                    |T $ %{} :Leaf (:at 1715484565342) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/sprott")
                     |b $ %{} :Leaf (:at 1709570997866) (:by |rJG4IHzWf) (:text |:as)
                     |e $ %{} :Leaf (:at 1709570999552) (:by |rJG4IHzWf) (:text |sprott)
                 |z8r $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1706460559814) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-lorenz83")
+                    |T $ %{} :Leaf (:at 1715484566836) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/lorenz83")
                     |b $ %{} :Leaf (:at 1709571002286) (:by |rJG4IHzWf) (:text |:as)
                     |e $ %{} :Leaf (:at 1709571008604) (:by |rJG4IHzWf) (:text |lorenz-83)
                 |z8v $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1710175465205) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-bouali")
+                    |T $ %{} :Leaf (:at 1715484568119) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/bouali")
                     |b $ %{} :Leaf (:at 1709571002286) (:by |rJG4IHzWf) (:text |:as)
                     |e $ %{} :Leaf (:at 1710175466701) (:by |rJG4IHzWf) (:text |bouali)
                 |z8x $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1714942296224) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-halvorsen")
+                    |T $ %{} :Leaf (:at 1715484569380) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/halvorsen")
                     |b $ %{} :Leaf (:at 1709571002286) (:by |rJG4IHzWf) (:text |:as)
                     |e $ %{} :Leaf (:at 1714942296919) (:by |rJG4IHzWf) (:text |halvorsen)
                 |z9 $ %{} :Expr (:at 1699548920367) (:by |rJG4IHzWf)
@@ -1403,7 +1403,7 @@
                     |h $ %{} :Leaf (:at 1709387579147) (:by |rJG4IHzWf) (:text |debug-grid)
                 |zCzY $ %{} :Expr (:at 1709356808034) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1710063641340) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-den-tsucs")
+                    |T $ %{} :Leaf (:at 1715484884808) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/den-tsucs")
                     |b $ %{} :Leaf (:at 1709356814139) (:by |rJG4IHzWf) (:text |:as)
                     |h $ %{} :Leaf (:at 1710063572855) (:by |rJG4IHzWf) (:text |den-tsucs)
                 |zCze $ %{} :Expr (:at 1709356808034) (:by |rJG4IHzWf)
@@ -1413,17 +1413,17 @@
                     |h $ %{} :Leaf (:at 1715106967436) (:by |rJG4IHzWf) (:text |clifford)
                 |zCzj $ %{} :Expr (:at 1709356808034) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1715197390360) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-dequanli")
+                    |T $ %{} :Leaf (:at 1715484888375) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/dequanli")
                     |b $ %{} :Leaf (:at 1709356814139) (:by |rJG4IHzWf) (:text |:as)
                     |h $ %{} :Leaf (:at 1715197291931) (:by |rJG4IHzWf) (:text |dequanli)
                 |zCzn $ %{} :Expr (:at 1709356808034) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1715355839777) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-dadras")
+                    |T $ %{} :Leaf (:at 1715484890295) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/dadras")
                     |b $ %{} :Leaf (:at 1709356814139) (:by |rJG4IHzWf) (:text |:as)
                     |h $ %{} :Leaf (:at 1715355842655) (:by |rJG4IHzWf) (:text |dadras)
                 |zCzq $ %{} :Expr (:at 1709356808034) (:by |rJG4IHzWf)
                   :data $ {}
-                    |T $ %{} :Leaf (:at 1715357743395) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor-burke-shaw")
+                    |T $ %{} :Leaf (:at 1715484891719) (:by |rJG4IHzWf) (:text "|\"../src/apps/attractor/burke-shaw")
                     |b $ %{} :Leaf (:at 1709356814139) (:by |rJG4IHzWf) (:text |:as)
                     |h $ %{} :Leaf (:at 1715357744406) (:by |rJG4IHzWf) (:text |burke-shaw)
                 |zD $ %{} :Expr (:at 1699549071462) (:by |rJG4IHzWf)
