+++ src/apps/attractor/aizawa.wgsl
