+++ src/apps/attractor/bouali.wgsl
