+++ src/apps/attractor/burke-shaw.wgsl
