+++ src/apps/attractor/chen.wgsl
