+++ src/apps/attractor/den-tsucs.mts
@@ -1,9 +1,9 @@
-import { createRenderer } from "../index.mjs";
-import attractorCompute from "./attractor-halvorsen.wgsl?raw";
-import { fiboGridN, rand_middle } from "../math.mjs";
+import { createRenderer } from "../../index.mjs";
+import attractorCompute from "./den-tsucs.wgsl?raw";
+import { fiboGridN, rand_middle } from "../../math.mjs";
 
 export let loadRenderer = async (canvas: HTMLCanvasElement) => {
-  let seedSize = 2000000;
+  let seedSize = 4000000;
 
   let renderFrame = await createRenderer(
     canvas,
@@ -11,9 +11,9 @@ export let loadRenderer = async (canvas: HTMLCanvasElement) => {
       seedSize,
       seedData: makeSeed(seedSize, 0),
       getParams: (dt) => [
-        dt * 0.04, // deltaT
-        600.0, // scale
-        0.001, // width
+        dt * 0.0003, // deltaT
+        20.0, // scale
+        0.008, // width
         0.99, // opacity
       ],
       computeShader: attractorCompute,
@@ -32,7 +32,7 @@ export let loadRenderer = async (canvas: HTMLCanvasElement) => {
 };
 
 let randPoint: [number, number, number] = [0, 0, 0];
-let area = 1.0;
+let area = 100.0;
 
 function makeSeed(numParticles: number, scale: number): Float32Array {
   const buf = new Float32Array(numParticles * 8);
@@ -42,15 +42,14 @@ function makeSeed(numParticles: number, scale: number): Float32Array {
       let p = fiboGridN(i, numParticles);
       randPoint = p;
     }
-
     let b = 8 * i;
-    buf[b + 0] = randPoint[0];
-    buf[b + 1] = randPoint[1];
-    buf[b + 2] = randPoint[2];
-    buf[b + 3] = rand_middle(0.8); // ages
-    buf[b + 4] = randPoint[0];
-    buf[b + 5] = randPoint[1];
-    buf[b + 6] = randPoint[2];
+    buf[b + 0] = randPoint[0] * area;
+    buf[b + 1] = randPoint[1] * area;
+    buf[b + 2] = randPoint[2] * area;
+    buf[b + 3] = i / 24; // ages
+    buf[b + 4] = randPoint[0] * area;
+    buf[b + 5] = randPoint[1] * area;
+    buf[b + 6] = randPoint[2] * area;
     buf[b + 7] = 0; // distance
   }
