+++ src/apps/attractor/den-tsucs.wgsl
