+++ src/apps/attractor/dequanli.mts
@@ -1,21 +1,24 @@
-import { createRenderer } from "../index.mjs";
-import attractorCompute from "./attractor-sprott.wgsl?raw";
-import { fiboGridN, rand_middle } from "../math.mjs";
+import { createRenderer } from "../../index.mjs";
+import attractorCompute from "./dequanli.wgsl?raw";
+import { fiboGridN, rand_middle } from "../../math.mjs";
 
 export let loadRenderer = async (canvas: HTMLCanvasElement) => {
-  let seedSize = 4000000;
+  // let seedSize = 2000000;
+  let seedSize = 2000000;
 
   let renderFrame = await createRenderer(
     canvas,
     {
       seedSize,
       seedData: makeSeed(seedSize, 0),
-      getParams: (dt) => [
-        dt * 0.04, // deltaT
-        600.0, // scale
-        0.001, // width
-        0.99, // opacity
-      ],
+      getParams: (dt) => {
+        return [
+          dt * 0.04, // deltaT
+          20.0, // scale
+          0.008, // width
+          0.99, // opacity
+        ];
+      },
       computeShader: attractorCompute,
     },
     {
@@ -32,22 +35,26 @@ export let loadRenderer = async (canvas: HTMLCanvasElement) => {
 };
 
 let randPoint: [number, number, number] = [0, 0, 0];
-let area = 1.0;
+let area = 80;
 
 function makeSeed(numParticles: number, scale: number): Float32Array {
   const buf = new Float32Array(numParticles * 8);
 
   for (let i = 0; i < numParticles; ++i) {
-    if (i % 24 == 0) {
-      let p = fiboGridN(i, numParticles);
-      randPoint = p;
+    if (i % 20 == 0) {
+      // let p = fiboGridN(i, numParticles);
+      randPoint = [
+        1.5 + rand_middle(area),
+        3.2 + rand_middle(area),
+        0.4 + rand_middle(area),
+      ];
     }
-
+    // console.log(randPoint);
     let b = 8 * i;
     buf[b + 0] = randPoint[0];
     buf[b + 1] = randPoint[1];
     buf[b + 2] = randPoint[2];
-    buf[b + 3] = rand_middle(0.8); // ages
+    buf[b + 3] = i / 20; // ages
     buf[b + 4] = randPoint[0];
     buf[b + 5] = randPoint[1];
     buf[b + 6] = randPoint[2];
