+++ src/apps/attractor/dequanli.wgsl
