+++ src/apps/attractor/fourwing.wgsl
