+++ src/apps/attractor/halvorsen.wgsl
