+++ src/apps/attractor/lorenz.wgsl
