+++ src/apps/attractor/lorenz83.mts
@@ -1,6 +1,6 @@
-import { createRenderer } from "../index.mjs";
-import attractorCompute from "./attractor-lorenz83.wgsl?raw";
-import { fiboGridN, rand_middle } from "../math.mjs";
+import { createRenderer } from "../../index.mjs";
+import attractorCompute from "./lorenz83.wgsl?raw";
+import { fiboGridN, rand_middle } from "../../math.mjs";
 
 export let loadRenderer = async (canvas: HTMLCanvasElement) => {
   let seedSize = 2000000;
