+++ src/apps/attractor/lorenz83.wgsl
