+++ src/apps/attractor/sprott.wgsl
